+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 **1.2.101**
 - Fixed bandwidth allocation for services.
 - Added callsign spotter to CW/RTTY skimmers.
+- Added CW spot reporting to PskReporter.
 - Added more checks to callsign-to-country conversion.
 - Added MQTT events for downloading data from the web.
 - Added timestamps to chat messages.